Split the Page class
